Matrixes double replaced with matrixes float, added world transform matrix

+++ Test/TestCamera.h
@@ -2,7 +2,7 @@
 #include "Eigen\Dense"
 #include <cmath>
 
-const double PI = 3.14159265358979323846;
+const float PI = 3.14159265358979323846;
 
 class TestCamera
 {
@@ -11,51 +11,51 @@ public:
 
     void SetProjectionValues(float fovDegrees, float aspectRatio, float nearZ, float farZ);
 
-    const Eigen::Matrix4d& GetViewMatrix() const;
-    const Eigen::Matrix4d& GetProjectionMatrix() const;
+    const Eigen::Matrix4f& GetViewMatrix() const;
+    const Eigen::Matrix4f& GetProjectionMatrix() const;
 
-    const Eigen::Vector3d& GetPosition() const;
-    const Eigen::Vector3d& GetRotation() const;
+    const Eigen::Vector3f& GetPosition() const;
+    const Eigen::Vector3f& GetRotation() const;
 
-    void SetPosition(const Eigen::Vector3d& pos);
-    void SetPosition(double x, double y, double z);
-    void AdjustPosition(const Eigen::Vector3d& pos);
-    void AdjustPosition(double x, double y, double z);
+    void SetPosition(const Eigen::Vector3f& pos);
+    void SetPosition(float x, float y, float z);
+    void AdjustPosition(const Eigen::Vector3f& pos);
+    void AdjustPosition(float x, float y, float z);
 
-    void MoveForward(double distance);
-    void MoveBackward(double distance);
-    void MoveRight(double distance);
-    void MoveLeft(double distance);
+    void MoveForward(float distance);
+    void MoveBackward(float distance);
+    void MoveRight(float distance);
+    void MoveLeft(float distance);
 
-    void SetRotation(const Eigen::Vector3d& rot);
-    void SetRotation(double pitch, double yaw, double roll);
-    void AdjustRotation(const Eigen::Vector3d& rot);
-    void AdjustRotation(double pitch, double yaw, double roll);
+    void SetRotation(const Eigen::Vector3f& rot);
+    void SetRotation(float pitch, float yaw, float roll);
+    void AdjustRotation(const Eigen::Vector3f& rot);
+    void AdjustRotation(float pitch, float yaw, float roll);
 
-    void SetLookAtPos(const Eigen::Vector3d& lookAtPos);
+    void SetLookAtPos(const Eigen::Vector3f& lookAtPos);
 
-    const Eigen::Vector3d& GetForwardVector();
-    const Eigen::Vector3d& GetRightVector();
-    const Eigen::Vector3d& GetBackwardVector();
-    const Eigen::Vector3d& GetLeftVector();
+    const Eigen::Vector3f& GetForwardVector();
+    const Eigen::Vector3f& GetRightVector();
+    const Eigen::Vector3f& GetBackwardVector();
+    const Eigen::Vector3f& GetLeftVector();
 
 private:
     void UpdateViewMatrix();
 
-    Eigen::Vector3d pos;
-    Eigen::Vector3d rot;
+    Eigen::Vector3f pos;
+    Eigen::Vector3f rot;
 
-    Eigen::Matrix4d viewMatrix;
-    Eigen::Matrix4d projectionMatrix;
+    Eigen::Matrix4f viewMatrix;
+    Eigen::Matrix4f projectionMatrix;
 
-    const Eigen::Vector3d DEFAULT_FORWARD_VECTOR = { 0.0, 0.0, 1.0 };
-    const Eigen::Vector3d DEFAULT_UP_VECTOR = { 0.0, 1.0, 0.0 };
-    const Eigen::Vector3d DEFAULT_BACKWARD_VECTOR = { 0.0, 0.0, -1.0 };
-    const Eigen::Vector3d DEFAULT_LEFT_VECTOR = { -1.0, 0.0, 0.0 };
-    const Eigen::Vector3d DEFAULT_RIGHT_VECTOR = { 1.0, 0.0, 0.0 };
+    const Eigen::Vector3f DEFAULT_FORWARD_VECTOR = { 0.0, 0.0, 1.0 };
+    const Eigen::Vector3f DEFAULT_UP_VECTOR = { 0.0, 1.0, 0.0 };
+    const Eigen::Vector3f DEFAULT_BACKWARD_VECTOR = { 0.0, 0.0, -1.0 };
+    const Eigen::Vector3f DEFAULT_LEFT_VECTOR = { -1.0, 0.0, 0.0 };
+    const Eigen::Vector3f DEFAULT_RIGHT_VECTOR = { 1.0, 0.0, 0.0 };
 
-    Eigen::Vector3d vec_forward;
-    Eigen::Vector3d vec_left;
-    Eigen::Vector3d vec_right;
-    Eigen::Vector3d vec_backward;
+    Eigen::Vector3f vec_forward;
+    Eigen::Vector3f vec_left;
+    Eigen::Vector3f vec_right;
+    Eigen::Vector3f vec_backward;
 };
